flask/requirements.txt
openai
pymongo
flask
bcrypt
python-decouple